Move `_ProcessWrapper` to different file
